+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0 (Unreleased)
 ------------------
 
+- #35 Rejection inbound samples
 - #32 Fix build test after adding default contact
 - #31 Add default contact for ExternalLaboratory and InboundSampleShipment content
 - First version